Flip the switch rman_res_t from u_long -> uintmax_t.
On some architectures u_long isn't large enough to handle some resource
definitions.  Particularly powerpc, arm, and i386 (32-bit platforms) can use
36-bit physical addresses, but the size of u_long is only 32 bits.  This removes
that limitation, at the expense of a small increase in code and data size.

In the future, it may be desirable to make rman_res_t a platform specific
typedef, so that platforms which don't require an exta large resource range
don't take the hit.  This is left for a future exercise.