libc: Add fdwalk(3), commonly used in Linux software
fdwalk(3) originated in SunOS/Solaris, and is commonly used in Linux
software, including GNOME-based software, like glib and vte.  It walks
the current file descriptor list and runs an arbitrary function on each
file descriptor.

To support this, add a 'kern.proc.fdmap.<proc>' sysctl, which provides
the file descriptor map for the process.